+++ Flutter/example/lib/common/demo_config.dart
@@ -4,8 +4,8 @@
 import 'dart:async';
 
 // The obtained license URL
-const LICENSE_URL = "";
+const LICENSE_URL = "https://license-test.vod2.myqcloud.com/test/license/testV2/1253968938_1/v_cube.license";
 // The obtained license key
-const LICENSE_KEY = "";
+const LICENSE_KEY = "79aa077679db4b0d68643e7af3516719";
 
 Completer<bool> isLicenseSuc = Completer();